Added configurable BOT_EXTRA_PLUGIN_DIR
This makes it possible to create a Docker image with rroemhild/errbot
as the base image, where additional plugins are copied in and used.

This is currenty not possible since BOT_EXTRA_PLUGIN_DIR is hardcoded
to /srv/plugins, /srv is a volume, and it is not possible to add files
to a volume after it has been created.